COform/requirements.txt
aniso8601==8.1.0
certifi==2020.12.5
click==7.1.2
config==0.5.0.post0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.3
PyJWT==2.0.1
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
wincertstore==0.2
